Refactor binConv.h
In this commit, the stdlib.h standary library function, "malloc" was
replaced with the "calloc" function, another stdlib.h standard library
function. The latter share some similarities and differences with the
former. There is no specific technical reason for this change as at this
snapshot of the project as either of the standard functions will work
well. However, the change was effected to avoid the default garbage
initialization of the "malloc" function in order to enhance future
maintenability.